+++ src/ts/components/dates/DateTimePicker.tsx
@@ -7,11 +7,10 @@ import {
     CalendarBaseProps,
     CalendarSettings,
     DateInputSharedProps,
-    TimeInputProps,
 } from "props/dates";
 import { StylesApiProps } from "props/styles";
 import React, { useState } from "react";
-import { datetimeToString, isDisabled, stringToDate } from "../../utils/dates";
+import { isDisabled } from "../../utils/dates";
 import { setPersistence, getLoadingState } from "../../utils/dash3";
 
 interface Props
@@ -29,8 +28,8 @@ interface Props
     valueFormat?: string;
     /** Controlled component value */
     value?: string;
-    /** TimeInput component props */
-    timeInputProps?: TimeInputProps;
+    /** Props passed the TimePicker component */
+    timePickerProps?: object;
     /** Props passed down to the submit button */
     submitButtonProps?: Omit<ActionIconProps, "n_click"> & any;
     /** Determines whether seconds input should be rendered */
@@ -61,18 +60,18 @@ const DateTimePicker = ({
     ...others
 }: Props) => {
 
-    const [date, setDate] = useState(stringToDate(value));
+    const [date, setDate] = useState(value);
     const [debounced] = useDebouncedValue(date, debounce);
 
     useDidUpdate(() => {
-        setProps({ value: datetimeToString(date) });
+        setProps({ value: date });
     }, [debounced]);
 
     useDidUpdate(() => {
-        setDate(stringToDate(value));
+        setDate(value);
     }, [value]);
 
-    const isExcluded = (date: Date) => {
+    const isExcluded = (date: string) => {
         return isDisabled(date, disabledDates || []);
     };
 
@@ -88,8 +87,8 @@ const DateTimePicker = ({
             onChange={setDate}
             value={date}
             onKeyDown={handleKeyDown}
-            minDate={stringToDate(minDate)}
-            maxDate={stringToDate(maxDate)}
+            minDate={minDate}
+            maxDate={maxDate}
             excludeDate={isExcluded}
             {...others}
         />