+++ src/ts/components/dates/MonthPickerInput.tsx
@@ -5,12 +5,6 @@ import { DashBaseProps, PersistenceProps } from "props/dash";
 import { DateInputSharedProps, MonthPickerBaseProps } from "props/dates";
 import { StylesApiProps } from "props/styles";
 import React, { useState } from "react";
-import {
-    isDisabled,
-    stringToDate,
-    toDates,
-    toStrings,
-} from "../../utils/dates";
 import { setPersistence, getLoadingState } from "../../utils/dash3";
 
 interface Props
@@ -48,26 +42,26 @@ const MonthPickerInput = ({
     ...others
 }: Props) => {
 
-    const [date, setDate] = useState(toDates(value));
+    const [date, setDate] = useState(value);
     const debounceValue = typeof debounce === 'number' ? debounce : 0;
     const [debounced] = useDebouncedValue(date, debounceValue);
     const { ref, focused } = useFocusWithin();
 
     useDidUpdate(() => {
         if (typeof debounce === 'number' || debounce === false) {
-            setProps({ value: toStrings(date) })
+            setProps({ value: date })
         };
     }, [debounced]);
 
     useDidUpdate(() => {
         // Clears value when X is clicked
         if (focused) {
-            setProps({ value: toStrings(date) });
+            setProps({ value: date });
         }
     }, [date]);
 
     useDidUpdate(() => {
-        setDate(type !== 'default' && !value ? [] : toDates(value));
+        setDate(type !== 'default' && !value ? [] : value);
     }, [value]);
 
     const handleKeyDown = (ev) => {
@@ -79,13 +73,10 @@ const MonthPickerInput = ({
     const handleBlur = () => {
         // Don't include n_blur counter because onBlur is called when the calendar is opened
         if (debounce === true) {
-            setProps({ value: toStrings(date) });
+            setProps({ value: date });
         }
     };
 
-    const isExcluded = (date: Date) => {
-        return isDisabled(date, disabledDates || []);
-    };
 
     return (
         <div ref={ref}>
@@ -96,8 +87,8 @@ const MonthPickerInput = ({
                 onChange={setDate}
                 value={date}
                 type={type}
-                minDate={stringToDate(minDate)}
-                maxDate={stringToDate(maxDate)}
+                minDate={minDate}
+                maxDate={maxDate}
                 popoverProps={{ returnFocus: true, ...popoverProps }}
                 {...others}
             />