+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+# Unreleased (2.0.0)
+
+### Changed
+ - BREAKING CHANGES - updated to use Mantine 8.0.2
+ - See [Migration guide](https://www.dash-mantine-components.com/migration) in the dmc-docs
+
+### Added
+  - New `TimePicker` and `TimeGrid`, `TimePicker` components
+
 # 1.3.0
 
 ### Added